chore: move common function to keeper_test.go
